feat(pos): учёт акций в корзине, онлайн и офлайн (#30)
* docs: add multi-seller PIN auth design spec

Several cashiers share one register and switch per receipt, but auth is a
single email+password login with no user identity in the app. Design adds
three identity levels: device token, shift login, and per-seller PIN
verified locally so switching works offline.


* docs: add multi-seller PIN auth implementation plan

Switches the PIN hash from argon2id to PBKDF2-HMAC-SHA256: the cash register
verifies PINs offline in C#, where argon2 needs a third-party package and
risks a cross-language mismatch. Both standard libraries ship PBKDF2.

Also records a discrepancy found while planning: IsCashSeller checks
is_seller while the roster query filters on can_sell, so body-supplied
sellers need a check requiring both flags.


* feat(pos): учёт акций в корзине, онлайн и офлайн

Бэкенд научился акциям (cloudmarket-server#32), но касса их не видела ни в
одном режиме. Четыре независимые причины:

- Котировка запрашивалась только при наличии карты или промокода. Акции
  применяются автоматически, входных данных у них нет, поэтому запрос не
  уходил никогда. Гейт снят: условие — онлайн, непустая корзина, известен
  склад.
- quote_id не отправлялся в документ расхода, из-за чего сервер не запускал
  FinalizeForSale: не было ни проверки дрейфа цен, ни списания источника.
- Ответ котировки не содержал названия акции, показать кассиру было нечего.
  Теперь имя источника видно в панели итогов и печатается в чеке.
- Офлайн акций не существовало вовсе. Добавлен локальный расчёт.

Офлайн-движок — порт discounts/source_promotion.go: лестница правил,
percent / amount / cheapest_free, best-deal по максимальной сумме. Тесты
повторяют серверные кейсы один в один, включая усечение весового товара до
целых единиц и запрет на сложение ступеней: расхождение расчётов означало бы
разную цену одной корзины онлайн и офлайн.

Набор акций и политика округления магазина кэшируются в SQLite и обновляются
вместе с товарами. Акции тянутся без фильтра по окну действия — касса может
простоять офлайн через момент старта или конца акции, поэтому окно она
проверяет своими часами. Ошибка запроса оставляет прежний кэш: считать по
вчерашним акциям лучше, чем не считать вовсе; пустой ответ, наоборот, кэш
чистит, иначе удалённая кампания продолжала бы давать скидку.

Офлайн-чек шлёт offline_promotion_id вместо quote_id, чтобы сервер всё же
списал использование акции — иначе max_uses игнорировался бы на каждой
продаже без связи. Акция рапортуется только когда выиграла у плоской скидки:
проигравшая ничего не дала и не должна тратить использование.

Заодно количество в позиции стало decimal. Сервер всегда принимал дробное,
обрезал только клиент, так что весовой товар продавался и списывался
округлённым до целого.


* docs: add cashier promotions implementation plan


---------

+++ src/VvCash/Models/ParkedSale.cs
@@ -9,7 +9,9 @@ public class ParkedSale
     public string? Label { get; set; }
     public string? CustomerName { get; set; }
     public decimal Total { get; set; }
-    public int ItemCount { get; set; }
+    /// <summary>Total units in the parked cart. Decimal because a weighted line
+    /// contributes a fraction of a unit.</summary>
+    public decimal ItemCount { get; set; }
     public DateTime CreatedAt { get; set; }
     public string Payload { get; set; } = string.Empty; // JSON ParkedSaleSnapshot
 }
